+++ assets/javascripts/kitten/components/form/dropdown-select/index.js
@@ -31,12 +31,14 @@ export const DropdownSelect = ({ combobox, ...props }) => {
     a11ySelectionMessageDisplayer,
     defaultSelectedValue,
     onChange,
+    onBlur,
     onInputChange,
     onMenuClose,
     onMenuOpen,
     openOnLoad,
     menuZIndex,
     className,
+    value,
   } = props
 
   const getA11ySelectionMessage = ({ itemToString, selectedItem }) => {
@@ -45,8 +47,6 @@ export const DropdownSelect = ({ combobox, ...props }) => {
 
   const itemToString = item => (item ? String(item.label) : '')
 
-  const initialSelectedItem = find(['value', defaultSelectedValue])(options)
-
   const onSelectedItemChange = changes => {
     onChange(changes.selectedItem)
     onInputChange({ value: changes.selectedItem })
@@ -54,7 +54,7 @@ export const DropdownSelect = ({ combobox, ...props }) => {
 
   // Turns a hierarchy of options with `children` into a flat array
   // of options with a `level` of 1, 2 or null.
-  const flattenedOptions = () => {
+  const flattenedOptions = (function () {
     const flatOptions = []
 
     options.map(option => {
@@ -70,13 +70,17 @@ export const DropdownSelect = ({ combobox, ...props }) => {
         flatOptions.push(option)
       }
     })
-
     return flatOptions
-  }
+  })()
+
+  const initialSelectedItem = find(['value', defaultSelectedValue])(
+    flattenedOptions,
+  )
+  const selectedItemByValue = find(['value', value])(flattenedOptions)
 
   const onIsOpenChange = changes => {
     if (changes.isOpen) return onMenuOpen({ changes })
-
+    setTimeout(() => onBlur(changes.selectedItem), 0)
     return onMenuClose({ changes })
   }
 
@@ -91,13 +95,14 @@ export const DropdownSelect = ({ combobox, ...props }) => {
   } = useSelect({
     id: `${id}_element`,
     toggleButtonId: id,
-    items: flattenedOptions(),
+    items: flattenedOptions,
     getA11ySelectionMessage,
     itemToString,
     initialSelectedItem,
     onSelectedItemChange,
     onIsOpenChange,
     initialIsOpen: openOnLoad,
+    ...(selectedItemByValue && { selectedItem: selectedItemByValue }),
   })
 
   useEffect(() => {
@@ -170,7 +175,7 @@ export const DropdownSelect = ({ combobox, ...props }) => {
       </button>
       <ul className="k-Form-Dropdown__list" {...getMenuProps()}>
         {isOpen &&
-          flattenedOptions().map((item, index) => (
+          flattenedOptions.map((item, index) => (
             <li
               className={classNames(
                 'k-Form-Dropdown__item',